Move content to new directories
